Add SetUpGame method to the MainWindow class
Add SetUpGame method, called from the MainWindow class. This method sets the
game up by creating a List of 8 pairs of animal emojis. Using a foreach
loop, the method assigns each item in the list to one of the text boxes in
the grid, removing each list item at the end of each iteration.