Add event handler to all TextBlock clicks and add game logic
TextBlock click handler contains the game logic. When a player first clicks
a tile, the tile is hidden and the value is stored, and a boolean is set from
false to true. If the player correctly matches on their second click, both
tiles remain hidden and the findingMatch boolean is set to false. If the
player incorrectly matches, the first tile is unhidden and findingMatch
is set to false.